requirements.txt
simpletransformers
torch
torchvision
torchaudio
opencv-python
matplotlib
git+https://github.com/facebookresearch/segment-anything.git
git+https://github.com/huggingface/transformers
pydantic
pydantic_settings
vision_transformers
diffusers
accelerate
scipy
safetensors
pyyaml
tqdm
numpy
easydict==1.9.0
scikit-image
scikit-learn
tensorflow
joblib
pandas
albumentations==0.5.2
hydra-core==1.1.0
tabulate
kornia==0.5.0
webdataset
packaging
wldhx.yadisk-direct
timm
pytorch-lightning
fastapi
uvicorn
python-multipart
streamlit
streamlit-aggrid
